Codec.Wavefront

Description

Currently, you can parse a file and get a WavefrontOBJ with the fromFile function.

Vertex location

data Location

A location is a 4-floating vector. You can access to its components by pattern matching on them:

  let Location x y z w = Location 1 2 3 4

That type is strict and unboxed.

Constructors

Location

Fields

locX :: !Float

locY :: !Float

locZ :: !Float

locW :: !Float

Vertex texture coordinates

data TexCoord

A texture coordinate is a 3D-floating vector. You can access to its components by pattern matching on them:

  let TexCoord r s t = TexCoord 0.1 0.2 0.3

That type is strict and unboxed.

Constructors

TexCoord

Fields

texcoordR :: !Float

texcoordS :: !Float

texcoordT :: !Float

Vertex normals

data Normal

A normal is a 3-floating vector. You can access to its components by pattern matching on them:

  let Normal nx ny nz = Normal 0.1 0.2 0.3

That type is strict and unboxed.

Constructors

Normal

Fields

norX :: !Float

norY :: !Float

norZ :: !Float

Points

data Point

A point is a single index that references the locations. It’s a canonical type that truly represents a polygonal point.

Constructors

Point

Fields

pointLocIndex :: !Int

Lines

data Line

Constructors

Line

Fields

lineIndexA :: LineIndex

lineIndexB :: LineIndex

data LineIndex

A line index is a pair of indices. LineIndex vi vti. vi references the locations and vti indexes the texture coordinates. If vti == Nothing, then that LineIndex doesn’t have texture coordinates associated with.

Constructors

LineIndex

Fields

lineLocIndex :: !Int

lineTexCoordIndex :: !(Maybe Int)

Faces

data Face

A face gathers several FaceIndex to build up faces. It has a least three vertices

Constructors

Face FaceIndex FaceIndex FaceIndex [FaceIndex]

data FaceIndex

A face index is a triplet of indices. FaceIndex vi vti vni is a face that indexes the locations with vi, the texture coordinates with vti and the normals with vni. An index set to Nothing means no information. That is, if vni == Nothing, then that FaceIndex doesn’t have a normal associated with.

Constructors

FaceIndex

Fields

faceLocIndex :: !Int

faceTexCoordIndex :: !(Maybe Int)

faceNorIndex :: !(Maybe Int)

pattern Triangle :: FaceIndex -> FaceIndex -> FaceIndex -> Face

pattern Quad :: FaceIndex -> FaceIndex -> FaceIndex -> FaceIndex -> Face

Element

type Element = ElementF (Maybe Material)

An element whose material has been dereferenced.

data ElementF material a

An element holds a value along with the user-defined object’s name (if any), the associated groups, the used material and the smoothing group the element belongs to (if any). Those values can be used to sort the data per object or per group and to lookup materials.

Constructors

Element

Fields

elObject :: Maybe Text

elGroups :: [Text]

elMtl :: material

elSmoothingGroup :: Natural

elValue :: a

type RawElement = ElementF (Maybe Text)

An element whose material has not been dereferenced yet

Object

type WavefrontOBJ = WavefrontOBJF MaterialLib (Maybe Material)

The content of a WavefrontOBJ file, in which the materials have been dereferenced.

data WavefrontOBJF mtlLib material

Prototype result for a WaveFrontOBJ File

Note according to the OBJ-spec, locations are "1-indexed"; i.e. in the range [1..n] rather than [0..n-1]. Hence, the indices in the Points, Lines, and Faces all refer to these 1-indexed points.

Constructors

WavefrontOBJ

Fields

objLocations :: Vector Location

Locations.

objTexCoords :: Vector TexCoord

Texture coordinates.

objNormals :: Vector Normal

Normals.

objPoints :: Vector (ElementF material Point)

Points.

objLines :: Vector (ElementF material Line)

Lines.

objFaces :: Vector (ElementF material Face)

Faces.

objMtlLibs :: Vector mtlLib

Material libraries.

type RawWavefrontOBJ = WavefrontOBJF FilePath (Maybe Text)

A wavefront object whose material has not been dereferenced yet

Material

data MaterialLib

A MaterialLib

data Material

Specification of a Material

Constructors

Material

Fields

materialName :: MaterialName

ambientReflexivity :: Maybe Reflexivity

diffuseReflexivity :: Maybe Reflexivity

specularReflexivity :: Maybe Reflexivity

emissiveReflexivity :: Maybe Reflexivity

transmissionFilter :: Maybe Reflexivity

iluminationModel :: Maybe IluminationModel

disolveFactor :: Maybe Float

specularExponent :: Maybe Float

sharpness :: Maybe Float

opticalDensity :: Maybe Float

Attributes of Materials

type MaterialName = Text

data RGB

Constructors

RGB !Float !Float !Float

data CIEXYZ

Constructors

CIEXYZ !Float !Float !Float

data Reflexivity

Constructors

ReflexicityRGB !RGB

ReflexivitySpectral FilePath (Maybe Float)

ReflexivityCIE !CIEXYZ

data IluminationModel

Constructors

ColorOnly

ColorAndAmbient

Higlight

ReflectionAndRayTrace

GlassAndRayTrace

FresnelAndRayTrace

RefractionAndRayTrace

RefactionFresnelAndRayTrace

ReflectionOnly

GlassOnly

CastsShadows

Re-exports

module Codec.Wavefront.IO