Codec.Wavefront.IO

Description

Documentation

fromFile :: MonadIO m => FilePath -> m (Either String WavefrontOBJ)

Extract a WavefrontOBJ from a Wavefront OBJ formatted file. If the file specifies mtl files, these are parsed and their information is parsed as well.

rawFromFile :: MonadIO m => FilePath -> m (Either String RawWavefrontOBJ)

Extract a raw WavefrontOBJ from a Wavefront OBJ formatted file. This only parses the given file, and does not dereference any materials.

fromText :: Text -> Either String RawWavefrontOBJ

Extract a WavefrontOBJ from an OBJ formatted text.

materialLibPath :: FilePath -> FilePath -> FilePath

given an obj file and its path to an mtl file; compute the absolute path

dereferenceMaterials :: RawWavefrontOBJ -> Vector MaterialLib -> Either String WavefrontOBJ

Given a raw OBJ Object, and a vector of MaterialLibs, tries to dereference all materials mentioned in the file.