Codec.Wavefront.Lexer

Description

Documentation

data Ctxt

The lexer context. The result of lexing a stream of tokens is this exact type.

Constructors

Ctxt

Fields

ctxtLocations :: SizedDList Location

Locations.

ctxtTexCoords :: SizedDList TexCoord

Texture coordinates.

ctxtNormals :: SizedDList Normal

Normals.

ctxtPoints :: DList (RawElement Point)

Points.

ctxtLines :: DList (RawElement Line)

Lines.

ctxtFaces :: DList (RawElement Face)

Faces.

ctxtCurrentObject :: Maybe Text

Current object.

ctxtCurrentGroups :: [Text]

Current groups.

ctxtCurrentMtl :: Maybe Text

Current material.

ctxtMtlLibs :: DList FilePath

Material libraries.

ctxtCurrentSmoothingGroup :: Natural

Current smoothing group.

emptyCtxt :: Ctxt

The empty Ctxt. Such a context exists at the beginning of the token stream and gets altered as we consume tokens.

lexer :: TokenStream -> Ctxt

The lexer function, consuming tokens and yielding a Ctxt.

prepareElement :: (Ctxt -> DList (RawElement a)) -> State Ctxt (DList (RawElement a), a -> RawElement a)

mapFaceIndices :: (FaceIndex -> FaceIndex) -> Face -> Face

map some function over the face indices

toAbsoluteIndex :: Int -> Int -> Int -> FaceIndex -> FaceIndex

Given the current number of locations, the current number of texture coordinates, and the current number of normals, make sure that the faceIndex uses absolute indices rather than relative ones.