Codec.Wavefront.Object

Description

Documentation

data WavefrontOBJF mtlLib material

Prototype result for a WaveFrontOBJ File

Note according to the OBJ-spec, locations are "1-indexed"; i.e. in the range [1..n] rather than [0..n-1]. Hence, the indices in the Points, Lines, and Faces all refer to these 1-indexed points.

Constructors

WavefrontOBJ

Fields

objLocations :: Vector Location

Locations.

objTexCoords :: Vector TexCoord

Texture coordinates.

objNormals :: Vector Normal

Normals.

objPoints :: Vector (ElementF material Point)

Points.

objLines :: Vector (ElementF material Line)

Lines.

objFaces :: Vector (ElementF material Face)

Faces.

objMtlLibs :: Vector mtlLib

Material libraries.

type WavefrontOBJ = WavefrontOBJF MaterialLib (Maybe Material)

The content of a WavefrontOBJ file, in which the materials have been dereferenced.

type RawWavefrontOBJ = WavefrontOBJF FilePath (Maybe Text)

A wavefront object whose material has not been dereferenced yet

ctxtToWavefrontOBJ :: Ctxt -> RawWavefrontOBJ

fromDList :: DList a -> Vector a

fromSizedDList :: SizedDList a -> Vector a