Codec.Wavefront.Element

Description

Element

data ElementF material a

An element holds a value along with the user-defined object’s name (if any), the associated groups, the used material and the smoothing group the element belongs to (if any). Those values can be used to sort the data per object or per group and to lookup materials.

Constructors

Element

Fields

elObject :: Maybe Text

elGroups :: [Text]

elMtl :: material

elSmoothingGroup :: Natural

elValue :: a

type Element = ElementF (Maybe Material)

An element whose material has been dereferenced.

type RawElement = ElementF (Maybe Text)

An element whose material has not been dereferenced yet