Codec.Wavefront.Face

Description

Documentation

data FaceIndex

A face index is a triplet of indices. FaceIndex vi vti vni is a face that indexes the locations with vi, the texture coordinates with vti and the normals with vni. An index set to Nothing means no information. That is, if vni == Nothing, then that FaceIndex doesn’t have a normal associated with.

Constructors

FaceIndex

Fields

faceLocIndex :: !Int

faceTexCoordIndex :: !(Maybe Int)

faceNorIndex :: !(Maybe Int)

data Face

A face gathers several FaceIndex to build up faces. It has a least three vertices

Constructors

Face FaceIndex FaceIndex FaceIndex [FaceIndex]

pattern Triangle :: FaceIndex -> FaceIndex -> FaceIndex -> Face

pattern Quad :: FaceIndex -> FaceIndex -> FaceIndex -> FaceIndex -> Face